requirements.txt
aiohttp==3.9.1
boto3==1.26.86
botocore==1.29.86
requests==2.31.0
simplejson==3.18.0
urllib3==1.26.14
watchtower==3.0.1
python-dotenv==1.0.1
